+++ psoc5/firmware/PSOC_trigger_firmware_Trigger_Polarity_Low/Trigger.cydsn/Generated_Source/PSoC5/cyfitter_cfg.c
@@ -148,7 +148,7 @@ static void CyClockStartupError(uint8 errorCode)
 }
 #endif
 
-#define CY_CFG_BASE_ADDR_COUNT 34u
+#define CY_CFG_BASE_ADDR_COUNT 35u
 CYPACKED typedef struct
 {
 	uint8 offset;
@@ -156,28 +156,28 @@ CYPACKED typedef struct
 } CYPACKED_ATTR cy_cfg_addrvalue_t;
 
 #define cy_cfg_addr_table ((const uint32 CYFAR *)0x48000000u)
-#define cy_cfg_data_table ((const cy_cfg_addrvalue_t CYFAR *)0x48000088u)
+#define cy_cfg_data_table ((const cy_cfg_addrvalue_t CYFAR *)0x4800008Cu)
 
 /* IOPORT_7 Address: CYDEV_PRTDSI_PRT12_BASE Size (bytes): 6 */
-#define BS_IOPORT_7_VAL ((const uint8 CYFAR *)0x4800066Cu)
+#define BS_IOPORT_7_VAL ((const uint8 CYFAR *)0x480006D4u)
 
 /* UDB_0_0_0_CONFIG Address: CYDEV_UCFG_B1_P5_U0_BASE Size (bytes): 128 */
-#define BS_UDB_0_0_0_CONFIG_VAL ((const uint8 CYFAR *)0x48000674u)
+#define BS_UDB_0_0_0_CONFIG_VAL ((const uint8 CYFAR *)0x480006DCu)
 
 /* IOPINS0_0 Address: CYREG_PRT0_DM0 Size (bytes): 8 */
-#define BS_IOPINS0_0_VAL ((const uint8 CYFAR *)0x480006F4u)
+#define BS_IOPINS0_0_VAL ((const uint8 CYFAR *)0x4800075Cu)
 
 /* IOPINS0_7 Address: CYREG_PRT12_DR Size (bytes): 10 */
-#define BS_IOPINS0_7_VAL ((const uint8 CYFAR *)0x480006FCu)
+#define BS_IOPINS0_7_VAL ((const uint8 CYFAR *)0x48000764u)
 
 /* IOPINS0_1 Address: CYREG_PRT1_DM0 Size (bytes): 8 */
-#define BS_IOPINS0_1_VAL ((const uint8 CYFAR *)0x48000708u)
+#define BS_IOPINS0_1_VAL ((const uint8 CYFAR *)0x48000770u)
 
 /* IOPINS0_2 Address: CYREG_PRT2_DR Size (bytes): 10 */
-#define BS_IOPINS0_2_VAL ((const uint8 CYFAR *)0x48000710u)
+#define BS_IOPINS0_2_VAL ((const uint8 CYFAR *)0x48000778u)
 
 /* IOPINS0_3 Address: CYREG_PRT3_DM0 Size (bytes): 8 */
-#define BS_IOPINS0_3_VAL ((const uint8 CYFAR *)0x4800071Cu)
+#define BS_IOPINS0_3_VAL ((const uint8 CYFAR *)0x48000784u)
 
 
 /*******************************************************************************