+++ psoc5/firmware/PSOC_trigger_firmware_Trigger_Polarity_Low/Trigger.cydsn/codegentemp/project.h
@@ -41,6 +41,7 @@
 #include "Timer.h"
 #include "isr_timer.h"
 #include "isr_swStart.h"
+#include "muxSel.h"
 #include "UART_IntClock.h"
 #include "core_cm3_psoc5.h"
 #include "CyDmac.h"